fixed key prop issue by adding a key prop to div containing input and label, so the parent element inside the return of the mapped array, not the input or label itself, also separated ArrowUp and ArrowDown functions into separate components and updated imports accordingly (was rendering an error after changing key prop), also added in margin-top for input, so that it is centered and in alignment with label
